pyproject.toml
[project]
name = "cold-protein-dti"
version = "0.1.0"
description = "Utilities and experiment scaffolding for cold-protein DTI research."
readme = "self/runtime_topology.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "numpy>=1.24,<3",
    "polars>=1.0,<2",
    "PyYAML>=6,<7",
]

[project.optional-dependencies]
train = [
    "scikit-learn>=1.4,<2",
    "torch>=2.1,<2.7",
]
viz = [
    "matplotlib>=3.7,<4",
    "umap-learn>=0.5,<0.6",
]
esm = [
    "fair-esm>=2.0.0,<3",
    "torch>=2.1,<2.7",
    "tqdm>=4.66,<5",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" },
]
